pazago_rag_assesment/pazago/requirements.txt
streamlit==1.28.1
openai==0.28.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
requests==2.31.0
sqlalchemy==2.0.23
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
tiktoken==0.5.1
chromadb==0.4.15